# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: CTWall Backend API
  version: 1.0.0
  description: |
    API for ChainThreatWall (CTWall) - SBOM Management and Security Analysis System.
    Adheres to OWASP ASVS Level 2 Security Requirements.
    Uses RFC 7807 for error handling.
servers:
  - url: https://api.ctwall.internal/api/v1
    description: Production Server
  - url: http://localhost:8080/api/v1
    description: Local Development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        API token for user/service accounts/CLI.
        Use: `Authorization: Bearer <API_TOKEN>`.
    CookieAuth:
      type: apiKey
      in: cookie
      name: __Host-CTWALL_SESSION
      description: |
        Session JWT stored in HttpOnly cookie for GUI users.
    RefreshCookieAuth:
      type: apiKey
      in: cookie
      name: __Secure-CTWALL_REFRESH
      description: |
        Refresh token stored in HttpOnly cookie for GUI session renewal.
  schemas:
...
    Error:
      type: object
      description: RFC 7807 Problem Details
      properties:
        type:
          type: string
          maxLength: 120
          format: uri
          example: "about:blank"
        title:
          type: string
          maxLength: 120
          example: "Invalid Request"
        status:
          type: integer
          example: 400
        detail:
          type: string
          maxLength: 120
          example: "Field 'sbom_file' is missing."
        instance:
          type: string
          maxLength: 120
          example: "/api/v1/ingest"
...
    Scope:
      type: object
      properties:
        id:
          type: string
          maxLength: 120
          format: uuid
        productId:
          type: string
          maxLength: 120
          format: uuid
        name:
          type: string
          maxLength: 120
        description:
          type: string
          maxLength: 120
        archivedAt:
          type: string
          maxLength: 120
          format: date-time
...
    SecurityPostureScore:
      type: object
      properties:
        value:
          type: integer
          minimum: 0
          maximum: 100
        label:
          type: string
          maxLength: 64
    SecurityPostureKpis:
      type: object
      properties:
        unknownPending:
          type: integer
        activeComponents:
          type: integer
        ingestFailures:
          type: integer
        ingestCompleted:
          type: integer
        ingestFailurePercent:
          type: number
          format: float
        osvSyncStatus:
          type: string
          maxLength: 64
        osvLastSyncAt:
          type: string
          maxLength: 120
          format: date-time
          nullable: true
        osvErrors:
          type: integer
        openErrorEvents:
          type: integer
        queueBacklog:
          type: integer
        inventoryTopType:
          type: string
          maxLength: 120
        lastAnalysisAt:
          type: string
          maxLength: 120
          format: date-time
          nullable: true
    SecurityPostureTopItem:
      type: object
      properties:
        name:
          type: string
          maxLength: 160
        value:
          type: integer
    SecurityPostureIngestBucket:
      type: object
      properties:
        bucketStart:
          type: string
          maxLength: 120
          format: date-time
        imports:
          type: integer
        failures:
          type: integer
    SecurityPostureSyncBucket:
      type: object
      properties:
        bucketStart:
          type: string
          maxLength: 120
          format: date-time
        runs:
          type: integer
        failures:
          type: integer
    SecurityPosturePackageType:
      type: object
      properties:
        packageType:
          type: string
          maxLength: 120
        count:
          type: integer
        percentage:
          type: number
          format: float
    SecurityPostureFailure:
      type: object
      properties:
        timestamp:
          type: string
          maxLength: 120
          format: date-time
        component:
          type: string
          maxLength: 240
        summary:
          type: string
          maxLength: 4000
        status:
          type: string
          maxLength: 64
    SecurityPostureUpload:
      type: object
      properties:
        id:
          type: string
          maxLength: 120
          format: uuid
        timestamp:
          type: string
          maxLength: 120
          format: date-time
        status:
          type: string
          maxLength: 64
        stage:
          type: string
          maxLength: 120
        componentsImported:
          type: integer
        productId:
          type: string
          maxLength: 120
          format: uuid
        scopeId:
          type: string
          maxLength: 120
          format: uuid
        testId:
          type: string
          maxLength: 120
          format: uuid
        productName:
          type: string
          maxLength: 240
        scopeName:
          type: string
          maxLength: 240
        testName:
          type: string
          maxLength: 240
        errorMessage:
          type: string
          maxLength: 4000
    SecurityPostureOverviewData:
      type: object
      properties:
        generatedAt:
          type: string
          maxLength: 120
          format: date-time
        rangeStart:
          type: string
          maxLength: 120
          format: date-time
        rangeEnd:
          type: string
          maxLength: 120
          format: date-time
        projectId:
          type: string
          maxLength: 120
          format: uuid
        topN:
          type: integer
        bucket:
          type: string
          maxLength: 16
          enum: [day, week]
        score:
          $ref: "#/components/schemas/SecurityPostureScore"
        kpis:
          $ref: "#/components/schemas/SecurityPostureKpis"
        ingestTrend:
          type: array
          items:
            $ref: "#/components/schemas/SecurityPostureIngestBucket"
        ingestFailureTop:
          type: array
          items:
            $ref: "#/components/schemas/SecurityPostureTopItem"
        osvSyncTrend:
          type: array
          items:
            $ref: "#/components/schemas/SecurityPostureSyncBucket"
        osvTopErrorStages:
          type: array
          items:
            $ref: "#/components/schemas/SecurityPostureTopItem"
        inventoryTopTypes:
          type: array
          items:
            $ref: "#/components/schemas/SecurityPosturePackageType"
        recentUploads:
          type: array
          items:
            $ref: "#/components/schemas/SecurityPostureUpload"
        recentFailures:
          type: array
          items:
            $ref: "#/components/schemas/SecurityPostureFailure"
    SecurityPostureOverviewResponse:
      type: object
      properties:
        scope:
          type: string
          maxLength: 32
          enum: [project]
        projectId:
          type: string
          maxLength: 120
          format: uuid
        range:
          type: string
          maxLength: 8
          enum: [24h, 7d, 30d, 90d]
        bucket:
          type: string
          maxLength: 16
          enum: [day, week]
        data:
          $ref: "#/components/schemas/SecurityPostureOverviewData"
...
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
...
  /security/posture/overview:
    get:
      summary: Project security posture overview (technical metrics)
      operationId: getSecurityPostureOverview
      parameters:
        - name: range
          in: query
          required: false
          schema:
            type: string
            maxLength: 8
            enum: [24h, 7d, 30d, 90d]
            default: 7d
        - name: topN
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 5
        - name: bucket
          in: query
          required: false
          schema:
            type: string
            maxLength: 16
            enum: [day, week]
            default: day
      responses:
        '200':
          description: Security posture overview loaded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityPostureOverviewResponse"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '500':
          description: Failed to load security posture overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # --- Events ---
...
  /products/{productId}/scopes:
    get:
      summary: List Scopes for Product
      operationId: listScopes
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            maxLength: 120
            format: uuid
      responses:
        '200':
          description: List of scopes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Scope'
    post:
      summary: Create Scope
      operationId: createScope
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            maxLength: 120
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 120
                description:
                  type: string
                  maxLength: 120
      responses:
        '201':
          description: Scope created
  /scopes/{scopeId}:
    delete:
      summary: Delete Scope
      operationId: deleteScope
      parameters:
        - name: scopeId
          in: path
          required: true
          schema:
            type: string
            maxLength: 120
            format: uuid
      responses:
        '204':
          description: Scope deleted
        '404':
          description: Scope not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # --- Tests ---
...